requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.27.0
aiofiles==23.2.1

# OCR and Document Processing
pytesseract==0.3.13
pdfplumber==0.11.0
Pillow==11.0.0

# AI/ML
openai==1.57.0

